requirements.txt
aiohttp
bs4
pyyaml
Pillow
Wand
markovify
psutil
imageio
numpy
molmass
si-prefix
redis
lxml
scipy
hiredis
matplotlib
xkcd
imagehash
uwuipy
py-cord[speed]==2.3.0
cchardet
brotlipy
aiodns
toml
aiofiles
pynacl
dulwich
wavelink
